Move some functions out of rustc_codegen_llvm and fix metadata_only backend
