+++ src/librustc_codegen_utils/lib.rs
@@ -47,6 +47,8 @@ pub mod link;
 pub mod codegen_backend;
 pub mod symbol_names;
 pub mod symbol_names_test;
+pub mod time_graph;
+pub mod llvm_target_features;
 
 /// check for the #[rustc_error] annotation, which forces an
 /// error in codegen. This is used to write compile-fail tests
