+++ src/librustc_codegen_utils/time_graph.rs
